requirements.txt
# Image processing and computer vision
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Data handling and configuration
scikit-image>=0.21.0

# Development and utilities
tqdm>=4.65.0  # For progress bars during batch generation 